requirements.txt
fastapi
uvicorn[standard]
fastapi-cache
sqlalchemy
pydantic
fastapi-cache2[memcache]
passlib
python-jose[cryptography]
python-multipart